# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MandalHisab API",
        "description": "Ganpati Mandal Financial Management System API. All responses use a standard envelope: `{success, statusCode, message, data, meta}`. Money amounts are INR with max 2 decimals. Calendar dates use `YYYY-MM-DD`, timestamps use ISO-8601 UTC.",
        "contact": {
            "email": "support@mandalhisab.in"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.mandalhisab.in/api/v1",
            "description": "Production"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development"
        }
    ],
    "paths": {
...
        "/festivals/{festivalId}/vargani": {
            "get": {
                "tags": [
                    "Vargani"
                ],
                "summary": "List vargani entries (paginated, filterable)",
                "operationId": "listVargani",
                "parameters": [
                    {
                        "name": "festivalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Donor name / phone / receipt #",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paymentMode",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "CASH",
                                "UPI",
                                "CHEQUE",
                                "NET_BANKING"
                            ]
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ALL",
                                "TODAY",
                                "CASH",
                                "UPI",
                                "CANCELLED"
                            ]
                        }
                    },
                    {
                        "name": "area",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collectorId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vargani entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VarganiListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vargani"
                ],
                "summary": "Create a vargani receipt (ADMIN/TREASURER/COLLECTOR)",
                "description": "Server assigns the next receipt number under a concurrency-safe lock. Sends Idempotency-Key recommended.",
                "operationId": "createVargani",
                "parameters": [
                    {
                        "name": "festivalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VarganiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receipt created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VarganiCreatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed (e.g. RECEIPT_BOOK_EXHAUSTED)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/festivals/{festivalId}/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "List expenses (paginated, filterable)",
                "operationId": "listExpenses",
                "parameters": [
                    {
                        "name": "festivalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "STAGE_MANDAP",
                                "SOUND_LIGHTING",
                                "MURTI_DECORATION",
                                "POOJA_PRASAD",
                                "SECURITY_LOGISTICS",
                                "MISCELLANEOUS"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PAID",
                                "PENDING"
                            ]
                        }
                    },
                    {
                        "name": "hasBill",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ALL",
                                "TODAY",
                                "PENDING",
                                "PENDING_BILL"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of expenses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpenseListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create an expense (ADMIN/TREASURER) — multipart or JSON",
                "description": "Either billFile OR billPendingReason is required. Idempotency-Key recommended.",
                "operationId": "createExpense",
                "parameters": [
                    {
                        "name": "festivalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpenseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VarganiRequest": {
                "required": [
                    "donorName",
                    "amount",
                    "paymentMode",
                    "area",
                    "receiptType"
                ],
                "properties": {
                    "donorName": {
                        "type": "string",
                        "maxLength": 60,
                        "minLength": 2
                    },
                    "mobileNumber": {
                        "description": "10 digits",
                        "type": "string",
                        "nullable": true
                    },
                    "amount": {
                        "type": "number",
                        "format": "double",
                        "exclusiveMinimum": true,
                        "minimum": 0
                    },
                    "paymentMode": {
                        "type": "string",
                        "enum": [
                            "CASH",
                            "UPI",
                            "CHEQUE",
                            "NET_BANKING"
                        ]
                    },
                    "area": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "collectorId": {
                        "description": "Server maps to collector",
                        "type": "string"
                    },
                    "receiptType": {
                        "type": "string",
                        "enum": [
                            "DIGITAL",
                            "PHYSICAL_BOOK"
                        ]
                    },
                    "receiptBookId": {
                        "description": "Required if PHYSICAL_BOOK",
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 250
                    },
                    "offlineSyncId": {
                        "description": "Client UUID for offline idempotency",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "statusCode": {
                        "type": "integer",
                        "example": 422
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "description": "One of: VALIDATION_FAILED, UNAUTHORIZED, TOKEN_EXPIRED, FORBIDDEN, NOT_FOUND, CONFLICT, ALREADY_CANCELLED, HANDOVER_NOT_PENDING, HISAB_LOCKED, RECEIPT_BOOK_EXHAUSTED, INVALID_PIN, BIOMETRIC_ATTESTATION_FAILED, FILE_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, RATE_LIMITED, MAINTENANCE_MODE, INTERNAL_ERROR",
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "details": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "field": {
                                            "type": "string"
                                        },
                                        "issue": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "VarganiListResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "statusCode": {
                        "type": "integer",
                        "example": 200
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VarganiItem"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
            "VarganiCreatedResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "statusCode": {
                        "type": "integer",
                        "example": 201
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "receiptNumber": {
                                "type": "string"
                            },
                            "donorName": {
                                "type": "string"
                            },
                            "amount": {
                                "type": "number",
                                "format": "double"
                            },
                            "paymentMode": {
                                "type": "string",
                                "enum": [
                                    "CASH",
                                    "UPI",
                                    "CHEQUE",
                                    "NET_BANKING"
                                ]
                            },
                            "isCancelled": {
                                "type": "boolean"
                            },
                            "receiptUrl": {
                                "type": "string"
                            },
                            "whatsappMessage": {
                                "type": "string"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token. Obtain by calling POST /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}